load GPS_point and save path into local file

+++ 6_QtApplication/minibrowser/gpspoint.h
@@ -8,38 +8,38 @@ class GpsPoint : public QObject
 {
     Q_OBJECT
 
-    Q_PROPERTY(float latitude READ getLatitude WRITE setLatitude NOTIFY latitudeChanged)
-    Q_PROPERTY(float longitude READ getLongitude WRITE setLongitude NOTIFY longitudeChanged)
+    Q_PROPERTY(double latitude READ getLatitude WRITE setLatitude NOTIFY latitudeChanged)
+    Q_PROPERTY(double longitude READ getLongitude WRITE setLongitude NOTIFY longitudeChanged)
 
-    float m_latitude;
-    float m_longitude;
+    double m_latitude;
+    double m_longitude;
 
 public:
     explicit GpsPoint(QObject *parent = nullptr);
-    GpsPoint (const float &latitude, const float &longtitude, QObject *parent = nullptr);
+    GpsPoint (const double &latitude, const double &longtitude, QObject *parent = nullptr);
     GpsPoint (const GpsPoint &aGpsPoint);
 
     bool userMoved(const GpsPoint &previousGpsPoint);
     void roundCoordinates();
-    float roundFloat (float &number);
+    double roundFloat (double &number);
 
-    float getLatitude() const
+    double getLatitude() const
     {
         return m_latitude;
     }
 
-    float getLongitude() const
+    double getLongitude() const
     {
         return m_longitude;
     }
 
 signals:
 
-    void latitudeChanged(float latitude);
-    void longitudeChanged(float longitude);
+    void latitudeChanged(double latitude);
+    void longitudeChanged(double longitude);
 
 public slots:
-    void setLatitude(float latitude)
+    void setLatitude(double latitude)
     {
         if (qFuzzyCompare(m_latitude, latitude))
             return;
@@ -47,7 +47,7 @@ public slots:
         m_latitude = latitude;
         emit latitudeChanged(m_latitude);
     }
-    void setLongitude(float longitude)
+    void setLongitude(double longitude)
     {
         if (qFuzzyCompare(m_longitude, longitude))
             return;
