+++ 6_QtApplication/minibrowser/mycontext.h
@@ -40,7 +40,7 @@ public:
     QString truncateUrl(const QString &url);
 
     Q_INVOKABLE void sendActionToCpp (QString nomAction, QString parameter = QString (""), QString parameter2 = QString ("") );
-    Q_INVOKABLE void updateTrek (QString actionType, float const &latitude, float const &longitude);
+    Q_INVOKABLE void updateTrek (QString actionType, double const &latitude, double const &longitude);
 
 
     Trek* getMyTrek() const
